requirements.txt
# Scouter
elasticsearch==6.5.1

# Document graph analyzer
numpy
gensim
python-igraph